Seperate proto-specific function adapter code from core implementation. Update expr:runtime::FunctionAdapter to be an alias of the proto-message enabled helper code.
PiperOrigin-RevId: 438933702
